Separate typescript functions into multiple files and add documentation
